[#9] refactor: move github package
from domain package to global package
